#7 feat: 예외처리

+++ src/main/java/com/example/myrok/service/MemberRecordServiceImpl.java
@@ -1,14 +1,17 @@
 package com.example.myrok.service;
 
-import com.example.myrok.domain.Member;
-import com.example.myrok.domain.MemberRecord;
+import com.example.myrok.domain.*;
 import com.example.myrok.domain.Record;
-import com.example.myrok.domain.RecordTag;
+import com.example.myrok.exception.CustomException;
 import com.example.myrok.exception.NotFoundException;
+import com.example.myrok.repository.MemberProjectRepository;
 import com.example.myrok.repository.MemberRecordRepository;
 import com.example.myrok.repository.MemberRepository;
+import com.example.myrok.type.ErrorCode;
 import com.example.myrok.type.Role;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -19,13 +22,22 @@ public class MemberRecordServiceImpl implements MemberRecordService{
     @Autowired
     MemberRecordRepository memberRecordRepository;
     @Autowired
+    MemberProjectRepository memberProjectRepository;
+    @Autowired
     MemberRepository memberRepository;
     @Override
     public List<MemberRecord> save(List<Long> members, Record record, Long recordWriterId){
         List<MemberRecord> memberRecordList= new ArrayList<>();
         for (Long memberId : members) {
             Member member = memberRepository.findById(memberId)
-                    .orElseThrow(() -> new NotFoundException("존재하지 않는 팀원입니다."));
+                    .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 멤버입니다. id: " + memberId));
+            // 멤버가 프로젝트 소속인지 확인
+            MemberProject memberProject=memberProjectRepository.findByMember(member)
+                    .orElseThrow(() -> new EntityNotFoundException("소속된 프로젝트가 없는 멤버입니다. id: " + memberId));
+            if(!record.getProject().getId().equals(memberProject.getProject().getId())){
+                throw new CustomException(ErrorCode.MEMBER_NOT_IN_PROJECT, HttpStatus.BAD_REQUEST);
+            }
+            // 멤버별 권한 부여
             Role role = memberId.equals(recordWriterId) ? Role.ADMIN : Role.PARTICIPANT;
             MemberRecord memberRecord = MemberRecord.builder()
                     .record(record)
@@ -41,6 +53,9 @@ public class MemberRecordServiceImpl implements MemberRecordService{
     public void delete(Long id){
         List<MemberRecord> memberRecords = memberRecordRepository.findAllByRecordId(id);
         for (MemberRecord memberRecord : memberRecords) {
+            if (memberRecord.getDeleted()){
+                throw new CustomException(ErrorCode.DELETED_MR_CODE, HttpStatus.BAD_REQUEST);
+            }
             memberRecord.delete();
             memberRecordRepository.save(memberRecord);
         }
