+++ src/main/java/com/example/myrok/service/TagServiceImpl.java
@@ -2,8 +2,11 @@ package com.example.myrok.service;
 
 import com.example.myrok.domain.RecordTag;
 import com.example.myrok.domain.Tag;
+import com.example.myrok.exception.CustomException;
 import com.example.myrok.repository.TagRepository;
+import com.example.myrok.type.ErrorCode;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,6 +16,8 @@ import java.util.List;
 public class TagServiceImpl implements TagService{
     @Autowired
     private TagRepository tagRepository;
+
+    @Transactional(rollbackFor = Exception.class)
     @Override
     public Tag save(String tagName){
         Tag tag = tagRepository.findByTagName(tagName)
@@ -23,12 +28,16 @@ public class TagServiceImpl implements TagService{
         return tag;
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
     public void delete(List<RecordTag> recordTagList){
         //Tag Count 감소, 0이라면 삭제
         //deleted 여부도 검사해야될듯
         for (RecordTag recordTag : recordTagList) {
             Tag tag= recordTag.getTag();
+            if (tag.getDeleted()) {
+                throw new CustomException(ErrorCode.DELETED_TAG_CODE, HttpStatus.BAD_REQUEST);
+            }
             tag.decrementCount();
             if (tag.getCount() <= 0) {
                 tag.delete();
